Fix up various nits
